scripts/requirements.txt
pandas
pdfplumber
pdfminer.six
requests
beautifulsoup4
rapidfuzz
playwright
